Renamed FeedItem to Comment
