+++ src/modules/core/components/Comment/Comment.css.d.ts
