+++ src/modules/core/components/Comment/Comment.tsx
@@ -11,9 +11,9 @@ import FriendlyName from '~core/FriendlyName';
 import { useDataFetcher } from '~utils/hooks';
 import { ipfsDataFetcher } from '../../fetchers';
 
-import styles from './FeedItem.css';
+import styles from './Comment.css';
 
-const displayName = 'FeedItem';
+const displayName = 'Comment';
 
 interface Props {
   comment?: string;
@@ -24,7 +24,7 @@ interface Props {
 
 const UserAvatar = HookedUserAvatar({ fetchUser: false });
 
-const FeedItem = ({ comment, user, createdAt, annotation = false }: Props) => {
+const Comment = ({ comment, user, createdAt, annotation = false }: Props) => {
   const { Decorate } = new TextDecorator({
     username: (usernameWithAtSign) => (
       <UserMention username={usernameWithAtSign.slice(1)} />
@@ -83,6 +83,6 @@ const FeedItem = ({ comment, user, createdAt, annotation = false }: Props) => {
   );
 };
 
-FeedItem.displayName = displayName;
+Comment.displayName = displayName;
 
-export default FeedItem;
+export default Comment;
