+++ src/modules/core/components/Comment/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Comment';
