--- src/modules/core/components/FeedItem/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FeedItem';
