+++ src/modules/dashboard/components/ActionsPageFeed/ActionsPageFeed.tsx
@@ -2,7 +2,7 @@ import React, { useState, useMemo, ReactElement, ReactNode } from 'react';
 import { nanoid } from 'nanoid';
 import { FormattedMessage, defineMessages } from 'react-intl';
 
-import FeedItem from '~core/FeedItem';
+import Comment from '~core/Comment';
 import { SpinnerLoader } from '~core/Preloaders';
 
 import { getEventsForActions } from '~utils/events';
@@ -214,7 +214,7 @@ const ActionsPageFeed = ({
           uniqueId,
         } = (feedItem as unknown) as FeedItemWithId<TransactionMessageFragment>;
         return (
-          <FeedItem
+          <Comment
             key={uniqueId}
             createdAt={createdAt}
             comment={message}
